Update README.md
Target assessment level
The target assessment level of this work is 2.

Specification
What does this program do?
The program
1.	Reads data user inputs
2.	Prints the dates when the user has agreed to do tasks suggested by the program

The user supplies only their name and a day they agree to do the assigned work. This is from the keyboard.
Data format
The input data consist of strings (name and day). Name can be as many words as the user wants to input, however, day should only be one word long.
Most of the data is however in char format, which creates the calendar background. 

Correctness and exception handling
Typical test case
The user gives the program their name, which after the program asks the user to give a date when they can do the required task.
Output is the calendar format.
In this case, typical input could be:
Name: Harri
Day: monday

Exception handling
The possible exceptions are:
•	User supplies data, which does not correspond to the required format. For example, the user supplies a date written in upper case. -> the program catches exceptions and gives an exception code 40.